mrc-misc-fn-modules(extract-expert-location-src-rank-move,mechanical_provable): move get_healthy_expert_location_src_rank to elastic_ep module (cut+paste)
PR-Title: Extract small single-function helpers into modules
