+++ python/sglang/srt/elastic_ep/elastic_ep.py
@@ -7,7 +7,7 @@ from typing import Iterator, List, Optional
 
 import torch
 
-from sglang.srt.distributed import parallel_state
+from sglang.srt.distributed import get_world_group, parallel_state
 from sglang.srt.managers.schedule_batch import ServerArgs
 from sglang.srt.utils import is_cpu, is_cuda
 
@@ -201,3 +201,23 @@ def join_process_groups():
         _maybe_create_message_queue(group)
 
     _refresh_ep_members()
+
+
+def get_healthy_expert_location_src_rank(
+    *, invoked_in_elastic_ep_rejoin_path: bool
+) -> int:
+    world_group = get_world_group()
+    # NOTE: do not key off `self.server_args.elastic_ep_rejoin` here.
+    # A rank that was started as a rejoin rank may later act as a healthy
+    # rank in a subsequent recovery cycle.
+    local_rejoin_flag = bool(invoked_in_elastic_ep_rejoin_path)
+    gathered_rejoin_flags = world_group.all_gather_object(local_rejoin_flag)
+
+    for rank_in_group, is_rejoin_rank in enumerate(gathered_rejoin_flags):
+        if not is_rejoin_rank:
+            return world_group.ranks[rank_in_group]
+
+    raise RuntimeError(
+        "No healthy rank found for broadcasting expert location metadata. "
+        "All ranks are marked as elastic_ep_rejoin."
+    )
