+++ python/sglang/srt/model_executor/model_runner.py
@@ -44,6 +44,7 @@ from sglang.srt.distributed.device_communicators.pynccl_allocator import (
 from sglang.srt.dllm.config import DllmConfig
 from sglang.srt.elastic_ep.elastic_ep import (
     ElasticEPStateManager,
+    get_healthy_expert_location_src_rank,
     join_process_groups,
     try_recover_ranks,
 )
@@ -1996,23 +1997,3 @@ class ModelRunner(ModelRunnerKVCacheMixin):
             load_format=load_format,
         )
         self.load_config = load_config
-
-
-def get_healthy_expert_location_src_rank(
-    *, invoked_in_elastic_ep_rejoin_path: bool
-) -> int:
-    world_group = get_world_group()
-    # NOTE: do not key off `self.server_args.elastic_ep_rejoin` here.
-    # A rank that was started as a rejoin rank may later act as a healthy
-    # rank in a subsequent recovery cycle.
-    local_rejoin_flag = bool(invoked_in_elastic_ep_rejoin_path)
-    gathered_rejoin_flags = world_group.all_gather_object(local_rejoin_flag)
-
-    for rank_in_group, is_rejoin_rank in enumerate(gathered_rejoin_flags):
-        if not is_rejoin_rank:
-            return world_group.ranks[rank_in_group]
-
-    raise RuntimeError(
-        "No healthy rank found for broadcasting expert location metadata. "
-        "All ranks are marked as elastic_ep_rejoin."
-    )
